Update README.md
Academic integrity is taken very seriously by the University. As this assignment is worth more than 10% of your course grade, any allegations of academic dishonesty go immediately to the Dean, and may take a long time to resolve. This is true whether you copied the work or ideas of someone else, or allowed someone to look at and copy your work.

Only show your work to your assignment partner or the course teaching team, and do not look at the work of anyone else.

If you have questions or need help, use the resources provided in the course.

Learning Goals

By the end of this assignment you will be able to:

read and interpret a novel schema written in SQL

write complex queries in SQL

design datasets to test a SQL query thoroughly